obj: fix a build failure with -fsanitize=undefined
This warning doesn't trigger otherwise.

+++ src/libpmemobj/critnib.c
@@ -207,7 +207,7 @@ path_mask(sh_t shift)
 static inline int
 slice_index(uint64_t key, sh_t shift)
 {
-	return (key >> shift) & NIB;
+	return (int)((key >> shift) & NIB);
 }
 
 /*
